# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me/referral:
    get:
      operationId: getMyReferral
      tags: [referrals]
      summary: Get (or lazily create) the current user's referral code, stats, and history
      responses:
        "200":
          description: Referral summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyReferralSummary"
        "401":
          description: Unauthorized
  /users/me/credit:
    get:
      operationId: getMyCredit
      tags: [referrals]
      summary: Get the current user's store credit balance
      responses:
        "200":
          description: Credit balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditBalance"
        "401":
          description: Unauthorized
...
  /admin/referral-settings:
    get:
      operationId: getReferralSettingsAdmin
      tags: [referrals]
      summary: Get referral programme configuration (admin)
      responses:
        "200":
          description: Referral settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminReferralSettings"
    put:
      operationId: updateReferralSettingsAdmin
      tags: [referrals]
      summary: Update referral programme configuration (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminReferralSettings"
        "400":
          description: Validation error
...
components:
  schemas:
...
    AdminReferralSettings:
      type: object
      required: [id, enabled, rewardType, rewardValue, minPurchaseKobo, updatedAt]
      properties:
        id:
          type: integer
        enabled:
          type: boolean
        rewardType:
          type: string
          description: "percentage | fixed | store_credit | free_product"
        rewardValue:
          type: integer
        rewardProductId:
          type: ["integer", "null"]
        minPurchaseKobo:
          type: integer
        campaignStartsAt:
          type: ["string", "null"]
        campaignEndsAt:
          type: ["string", "null"]
        maxRewardsPerReferrer:
          type: ["integer", "null"]
        updatedAt:
          type: string
        updatedBy:
          type: ["string", "null"]
    ReferralSettingsInput:
      type: object
      properties:
        enabled:
          type: boolean
        rewardType:
          type: string
        rewardValue:
          type: integer
        rewardProductId:
          type: ["integer", "null"]
        minPurchaseKobo:
          type: integer
        campaignStartsAt:
          type: ["string", "null"]
        campaignEndsAt:
          type: ["string", "null"]
        maxRewardsPerReferrer:
          type: ["integer", "null"]
...
    MyReferralSummary:
      type: object
      required: [code, totalReferrals, completedReferrals, pendingReferrals, totalRewardedKobo, referrals]
      properties:
        code:
          type: string
        totalReferrals:
          type: integer
        completedReferrals:
          type: integer
        pendingReferrals:
          type: integer
        totalRewardedKobo:
          type: integer
        referrals:
          type: array
          items:
            type: object
            required: [id, status, createdAt]
            properties:
              id:
                type: integer
              status:
                type: string
              rewardKobo:
                type: ["integer", "null"]
              createdAt:
                type: string
              completedAt:
                type: ["string", "null"]
    CreditBalance:
      type: object
      required: [balanceKobo]
      properties:
        balanceKobo:
          type: integer
...