# lib/api-zod/src/generated/types/myReferralSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MyReferralSummaryReferralsItem } from './myReferralSummaryReferralsItem';

export interface MyReferralSummary {
  code: string;
  totalReferrals: number;
  completedReferrals: number;
  pendingReferrals: number;
  totalRewardedKobo: number;
  referrals: MyReferralSummaryReferralsItem[];
}