# lib/api-zod/src/generated/types/referralSettingsInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReferralSettingsInput {
  enabled?: boolean;
  rewardType?: string;
  rewardValue?: number;
  /** @nullable */
  rewardProductId?: number | null;
  minPurchaseKobo?: number;
  /** @nullable */
  campaignStartsAt?: string | null;
  /** @nullable */
  campaignEndsAt?: string | null;
  /** @nullable */
  maxRewardsPerReferrer?: number | null;
}